backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
langchain==0.0.340
openai==1.3.5
groq==0.4.2
httpx==0.25.1
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
